requirements.txt
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
idna==2.8
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.14.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
powerline-status==2.7
prompt-toolkit==1.0.16
ptyprocess==0.6.0
Pygments==2.4.2
PyYAML==5.1.1
requests==2.22.0
simplegeneric==0.8.1
six==1.12.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
